// src/app/widget/ubicar-mediante-coordenadas/interfaces/CoordenadaSexagecimal.ts
/**
 * Interface para los grados de tipo sexagecimal
 * @date 05-05-2025
 * @author Andres Fabian Simbaqueba Del Rio
 */
export interface CoordenadaSexagecimal {
  longitudeDegrees: number;
  longitudeMinutes: number;
  longitudeSeconds: number;
  latitudeDegrees: number;
  latitudeMinutes: number;
  latitudeSeconds: number;
}